Move utils + tests into separate dirs for tidyness
