+++ src/app/lib/utilities/deepGet/index.test.js
@@ -1,4 +1,4 @@
-import deepGet from './deepGet';
+import deepGet from '.';
 
 const fixtureData = {
   foo: 123,
